Moved `View` component to `view.tsx`
