+++ src/view.tsx
@@ -1,20 +1,42 @@
 import * as React from 'react';
-import { Todo } from './fetch';
+import { Todo, TrelloNotifications } from './fetch';
 
-//const todos = await ConfluenceInlineTasks();
-const todos: Array<Todo> = [];
-const todoItems = todos.map((todo) => {
-  return (
-    <li>{todo.title}</li>
-  );
-});
+interface ViewProps {
+  title: string;
+}
 
-class View extends React.Component<undefined, undefined> {
+interface ViewState {
+  todos: Array<Todo>;
+}
+
+class View extends React.Component<ViewProps, ViewState> {
+
+  constructor(props: ViewProps) {
+    super(props);
+    let todos: Array<Todo> = [];
+    this.state = {todos: todos};
+  }
+
+  componentDidMount() {
+    TrelloNotifications().then((todos) => {
+      this.setState({todos: todos});
+    });
+  }
+
+  /*componentWillUnmount() {
+  }*/
+
   render() {
+    const todoListItems = this.state.todos.map((todo) => {
+      return <li key={todo.created}>{todo.title}</li>;
+    });
     return (
-      <ul>{todoItems}</ul>
+      <div>
+        <h2>{this.props.title}</h2>
+        <ul>{todoListItems}</ul>
+      </div>
     );
   }
 }
 
-export { View };
+export { ViewProps, ViewState, View };
